+++ Engine/Image.cpp
@@ -964,6 +964,12 @@ Image::pasteFromForDepth(const Image & srcImg,
          src += srcRowElements,
          dst += dstRowElements) {
         std::memcpy(dst, src, roi.width() * sizeof(PIX) * _nbComponents);
+#ifdef DEBUG_NAN
+        for (int i = 0; i < roi.width() * _nbComponents; ++i) {
+            assert( !(boost::math::isnan)(src[i]) ); // check for NaN
+            assert( !(boost::math::isnan)(dst[i]) ); // check for NaN
+        }
+#endif
     }
 } // Image::pasteFromForDepth
 
@@ -1047,7 +1053,7 @@ Image::resizeInternal(const Image* srcImg,
         if ( !aRect.isNull() ) {
             char* pix = (char*)wacc.pixelAt(aRect.x1, aRect.y1);
             assert(pix);
-            double a = aRect.area();
+            U64 a = aRect.area();
             std::size_t memsize = a * pixelSize;
             std::memset(pix, 0, memsize);
             if ( setBitmapTo1 && (*outputImage)->usesBitMap() ) {
@@ -1059,7 +1065,7 @@ Image::resizeInternal(const Image* srcImg,
         if ( !cRect.isNull() ) {
             char* pix = (char*)wacc.pixelAt(cRect.x1, cRect.y1);
             assert(pix);
-            double a = cRect.area();
+            U64 a = cRect.area();
             std::size_t memsize = a * pixelSize;
             std::memset(pix, 0, memsize);
             if ( setBitmapTo1 && (*outputImage)->usesBitMap() ) {
@@ -1432,6 +1438,9 @@ Image::fillForDepthForComponents(const RectI & roi_,
         for (int j = 0; j < roi.width(); ++j, dst += nComps) {
             for (int k = 0; k < nComps; ++k) {
                 dst[k] = fillValue[k];
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(dst[k]) ); // check for NaN
+#endif
             }
         }
     }
@@ -1852,7 +1861,9 @@ Image::halveRoIForDepth(const RectI & roi,
     RectI dstRoI;
     RectI srcRoI = roi;
     srcRoI.intersect(srcBounds, &srcRoI); // intersect srcRoI with the region of definition
-
+#ifdef DEBUG
+    assert(!checkForNaNsNoLock(srcRoI));
+#endif
     dstRoI.x1 = (srcRoI.x1 + 1) / 2; // equivalent to ceil(srcRoI.x1/2.0)
     dstRoI.y1 = (srcRoI.y1 + 1) / 2; // equivalent to ceil(srcRoI.y1/2.0)
     dstRoI.x2 = srcRoI.x2 / 2; // equivalent to floor(srcRoI.x2/2.0)
@@ -1922,7 +1933,12 @@ Image::halveRoIForDepth(const RectI & roi,
                 const PIX b = (pickNextCol && pickThisRow) ? *(srcPixStart + k + _nbComponents) : 0;
                 const PIX c = (pickThisCol && pickNextRow) ? *(srcPixStart + k + srcRowSize) : 0;
                 const PIX d = (pickNextCol && pickNextRow) ? *(srcPixStart + k + srcRowSize  + _nbComponents)  : 0;
-
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(a) ); // check for NaN
+                assert( !(boost::math::isnan)(b) ); // check for NaN
+                assert( !(boost::math::isnan)(c) ); // check for NaN
+                assert( !(boost::math::isnan)(d) ); // check for NaN
+#endif
                 assert( sumW == 2 || ( sumW == 1 && ( (a == 0 && c == 0) || (b == 0 && d == 0) ) ) );
                 assert( sumH == 2 || ( sumH == 1 && ( (a == 0 && b == 0) || (c == 0 && d == 0) ) ) );
                 dstPixStart[k] = (a + b + c + d) / sum;
@@ -2022,7 +2038,15 @@ Image::halve1DImageForDepth(const RectI & roi,
         assert(src && dst);
         for (int x = 0; x < halfWidth; ++x) {
             for (int k = 0; k < _nbComponents; ++k) {
-                *dst++ = PIX( (float)( *src + *(src + _nbComponents) ) / 2. );
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(*src) ); // check for NaN
+                assert( !(boost::math::isnan)(*(src + _nbComponents)) ); // check for NaN
+#endif
+                *dst = PIX( (float)( *src + *(src + _nbComponents) ) / 2. );
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(*dst) ); // check for NaN
+#endif
+                ++dst;
                 ++src;
             }
             src += _nbComponents;
@@ -2034,7 +2058,15 @@ Image::halve1DImageForDepth(const RectI & roi,
         assert(src && dst);
         for (int y = 0; y < halfHeight; ++y) {
             for (int k = 0; k < _nbComponents; ++k) {
-                *dst++ = PIX( (float)( *src + (*src + rowSize) ) / 2. );
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(*src) ); // check for NaN
+                assert( !(boost::math::isnan)(*(src + rowSize)) ); // check for NaN
+#endif
+                *dst = PIX( (float)( *src + (*src + rowSize) ) / 2. );
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(*dst) ); // check for NaN
+#endif
+                ++dst;
                 ++src;
             }
             src += rowSize;
@@ -2106,7 +2138,7 @@ Image::downscaleMipMap(const RectD& dstRod,
 }
 
 bool
-Image::checkForNaNs(const RectI& roi)
+Image::checkForNaNsAndFix(const RectI& roi)
 {
     if (getBitDepth() != eImageBitDepthFloat) {
         return false;
@@ -2125,8 +2157,43 @@ Image::checkForNaNs(const RectI& roi)
         for (; pix < end; ++pix) {
             // we remove NaNs, but infinity values should pose no problem
             // (if they do, please explain here which ones)
+#ifdef DEBUG_NAN
+            assert( !(boost::math::isnan)(*pix) ); // check for NaN
+#endif
             if ( (boost::math::isnan)(*pix) ) { // check for NaN ((boost::math::isnan)(x) is not slower than x != x and works with -Ofast)
                 *pix = 1.;
+                hasnan = true;
+            }
+        }
+    }
+
+    return hasnan;
+}
+
+bool
+Image::checkForNaNsNoLock(const RectI& roi) const
+{
+    if (getBitDepth() != eImageBitDepthFloat) {
+        return false;
+    }
+    if (getStorageMode() == eStorageModeGLTex) {
+        return false;
+    }
+
+    //QWriteLocker k(&_entryLock);
+    unsigned int compsCount = getComponentsCount();
+    bool hasnan = false;
+    for (int y = roi.y1; y < roi.y2; ++y) {
+        float* pix = (float*)pixelAt(roi.x1, y);
+        float* const end = pix +  compsCount * roi.width();
+
+        for (; pix < end; ++pix) {
+            // we remove NaNs, but infinity values should pose no problem
+            // (if they do, please explain here which ones)
+#ifdef DEBUG_NAN
+            assert( !(boost::math::isnan)(*pix) ); // check for NaN
+#endif
+            if ( (boost::math::isnan)(*pix) ) { // check for NaN ((boost::math::isnan)(x) is not slower than x != x and works with -Ofast)
                 hasnan = true;
             }
         }
@@ -2203,14 +2270,20 @@ Image::upscaleMipMapForDepth(const RectI & roi,
                 assert( ( dstPix - (PIX*)output->pixelAt(dstRoi.x1, dstRoi.y1) ) % _nbComponents == 0 );
                 assert(dstPix >= (PIX*)output->pixelAt(xo, yo) && dstPix < (PIX*)output->pixelAt(xo, yo) + xcount * _nbComponents);
                 for (int c = 0; c < _nbComponents; ++c) {
+#ifdef DEBUG_NAN
+                    assert( !(boost::math::isnan)(srcPix[c]) ); // check for NaN
+#endif
                     dstPix[c] = srcPix[c];
                 }
             }
             //assert(dstPix == dstPixFirst + xcount*components);
         }
+        assert(dstLineBatchStart == (PIX*)output->pixelAt(dstRoi.x1, yo));
         PIX * dstLineStart = dstLineBatchStart + dstRowSize; // first line was filled already
         // now replicate the line as many times as necessary
         for (int i = 1; i < ycount; ++i, dstLineStart += dstRowSize) {
+            assert(dstLineStart == (PIX*)output->pixelAt(dstRoi.x1, yo+i));
+            assert(dstLineStart + dstRowSize == (PIX*)output->pixelAt(dstRoi.x2 - 1, yo+i) + _nbComponents);
             std::copy(dstLineBatchStart, dstLineBatchStart + dstRowSize, dstLineStart);
         }
     }
@@ -2343,6 +2416,7 @@ Bitmap::copyRowPortion(int x1,
     char* dstBitmap = getBitmapAt(x1, y);
     const char* end = dstBitmap + (x2 - x1);
 
+    assert(srcBitmap && dstBitmap);
     while (dstBitmap < end) {
         *dstBitmap = /**srcBitmap == PIXEL_UNAVAILABLE ? 0 : */ *srcBitmap;
         ++dstBitmap;
@@ -2398,7 +2472,13 @@ Image::premultInternal(const RectI& roi)
     PIX* dstPix = (PIX*)acc.pixelAt(renderWindow.x1, renderWindow.y1);
     for ( int y = renderWindow.y1; y < renderWindow.y2; ++y, dstPix += (srcRowElements - (renderWindow.x2 - renderWindow.x1) * 4) ) {
         for (int x = renderWindow.x1; x < renderWindow.x2; ++x, dstPix += 4) {
+#ifdef DEBUG_NAN
+            assert( !(boost::math::isnan)(dstPix[3]) ); // check for NaN
+#endif
             for (int c = 0; c < 3; ++c) {
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(dstPix[c]) ); // check for NaN
+#endif
                 if (doPremult) {
                     dstPix[c] = PIX(float(dstPix[c]) * dstPix[3]);
                 } else {
@@ -2406,6 +2486,9 @@ Image::premultInternal(const RectI& roi)
                         dstPix[c] = PIX( dstPix[c] / float(dstPix[3]) );
                     }
                 }
+#ifdef DEBUG_NAN
+                assert( !(boost::math::isnan)(dstPix[c]) ); // check for NaN
+#endif
             }
         }
     }